pyproject.toml
[tool.poetry]
name = "ResizeWizard"
version = "0.0.1"
description = ""
authors = ["chatflip <oo.chat.flip@gmail.com>"]
package-mode = false

[tool.poetry.dependencies]
python = "^3.10,<3.11"
flask = "2.1.0"
pillow = "8.4.0"
werkzeug = "2.0.3"

[tool.poetry.group.dev.dependencies]
ruff = "0.5.7"
mypy = "1.11.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
